+++ src/algorithms/mul.rs
@@ -1,6 +1,6 @@
 #![allow(clippy::module_name_repetitions)]
 
-use crate::algorithms::{borrowing_sub, DoubleWord};
+use crate::algorithms::{borrowing_sub, ConstDoubleWord as DW};
 
 /// ⚠️ Computes `result += a * b` and checks for overflow.
 #[doc = crate::algorithms::unstable_warning!()]
@@ -21,7 +21,7 @@ use crate::algorithms::{borrowing_sub, DoubleWord};
 /// assert_eq!(result, [12]);
 /// ```
 #[inline(always)]
-pub fn addmul(mut lhs: &mut [u64], mut a: &[u64], mut b: &[u64]) -> bool {
+pub const fn addmul(mut lhs: &mut [u64], mut a: &[u64], mut b: &[u64]) -> bool {
     // Trim zeros from `a`
     while let [0, rest @ ..] = a {
         a = rest;
@@ -53,8 +53,10 @@ pub fn addmul(mut lhs: &mut [u64], mut a: &[u64], mut b: &[u64]) -> bool {
     let (a, b) = if b.len() > a.len() { (b, a) } else { (a, b) };
 
     // Iterate over limbs of `b` and add partial products to `lhs`.
+    let mut i = 0;
     let mut overflow = false;
-    for &b in b {
+    while i < b.len() {
+        let b = b[i];
         if lhs.len() >= a.len() {
             let (target, rest) = lhs.split_at_mut(a.len());
             let carry = addmul_nx1(target, a, b);
@@ -65,9 +67,10 @@ pub fn addmul(mut lhs: &mut [u64], mut a: &[u64], mut b: &[u64]) -> bool {
             if lhs.is_empty() {
                 break;
             }
-            addmul_nx1(lhs, &a[..lhs.len()], b);
+            addmul_nx1(lhs, a.split_at(lhs.len()).0, b);
         }
-        lhs = &mut lhs[1..];
+        lhs = lhs.split_at_mut(1).1;
+        i += 1;
     }
     overflow
 }
@@ -79,7 +82,7 @@ const ADDMUL_N_SMALL_LIMIT: usize = 8;
 #[doc = crate::algorithms::unstable_warning!()]
 /// See [`addmul`] for more details.
 #[inline(always)]
-pub fn addmul_n(lhs: &mut [u64], a: &[u64], b: &[u64]) {
+pub const fn addmul_n(lhs: &mut [u64], a: &[u64], b: &[u64]) {
     let n = lhs.len();
     if n <= ADDMUL_N_SMALL_LIMIT && a.len() == n && b.len() == n {
         addmul_n_small(lhs, a, b);
@@ -89,32 +92,38 @@ pub fn addmul_n(lhs: &mut [u64], a: &[u64], b: &[u64]) {
 }
 
 #[inline(always)]
-fn addmul_n_small(lhs: &mut [u64], a: &[u64], b: &[u64]) {
+const fn addmul_n_small(lhs: &mut [u64], a: &[u64], b: &[u64]) {
     let n = lhs.len();
     assume!(n <= ADDMUL_N_SMALL_LIMIT);
     assume!(a.len() == n);
     assume!(b.len() == n);
 
-    for j in 0..n {
+    let mut j = 0;
+    while j < n {
         let mut carry = 0;
-        for i in 0..(n - j) {
-            (lhs[j + i], carry) = u128::muladd2(a[i], b[j], carry, lhs[j + i]).split();
+        let mut i = 0;
+        while i < (n - j) {
+            (lhs[j + i], carry) = DW::carrying_mul_add(a[i], b[j], carry, lhs[j + i]).split();
+            i += 1;
         }
+        j += 1;
     }
 }
 
 /// ⚠️ Computes `lhs += a` and returns the carry.
 #[doc = crate::algorithms::unstable_warning!()]
 #[inline(always)]
-pub fn add_nx1(lhs: &mut [u64], mut a: u64) -> u64 {
+pub const fn add_nx1(lhs: &mut [u64], mut a: u64) -> u64 {
     if a == 0 {
         return 0;
     }
-    for lhs in lhs {
-        (*lhs, a) = u128::add(*lhs, a).split();
+    let mut i = 0;
+    while i < lhs.len() {
+        (lhs[i], a) = DW::add(lhs[i], a).split();
         if a == 0 {
             return 0;
         }
+        i += 1;
     }
     a
 }
@@ -122,10 +131,12 @@ pub fn add_nx1(lhs: &mut [u64], mut a: u64) -> u64 {
 /// ⚠️ Computes `lhs *= a` and returns the carry.
 #[doc = crate::algorithms::unstable_warning!()]
 #[inline(always)]
-pub fn mul_nx1(lhs: &mut [u64], a: u64) -> u64 {
+pub const fn mul_nx1(lhs: &mut [u64], a: u64) -> u64 {
     let mut carry = 0;
-    for lhs in lhs {
-        (*lhs, carry) = u128::muladd(*lhs, a, carry).split();
+    let mut i = 0;
+    while i < lhs.len() {
+        (lhs[i], carry) = DW::carrying_mul(lhs[i], a, carry).split();
+        i += 1;
     }
     carry
 }
@@ -141,11 +152,13 @@ pub fn mul_nx1(lhs: &mut [u64], a: u64) -> u64 {
 /// }{2^{64⋅N}}} \end{aligned}
 /// $$
 #[inline(always)]
-pub fn addmul_nx1(lhs: &mut [u64], a: &[u64], b: u64) -> u64 {
+pub const fn addmul_nx1(lhs: &mut [u64], a: &[u64], b: u64) -> u64 {
     assume!(lhs.len() == a.len());
     let mut carry = 0;
-    for i in 0..a.len() {
-        (lhs[i], carry) = u128::muladd2(a[i], b, carry, lhs[i]).split();
+    let mut i = 0;
+    while i < a.len() {
+        (lhs[i], carry) = DW::carrying_mul_add(a[i], b, carry, lhs[i]).split();
+        i += 1;
     }
     carry
 }
@@ -162,17 +175,19 @@ pub fn addmul_nx1(lhs: &mut [u64], a: &[u64], b: u64) -> u64 {
 /// $$
 // OPT: `carry` and `borrow` can probably be merged into a single var.
 #[inline(always)]
-pub fn submul_nx1(lhs: &mut [u64], a: &[u64], b: u64) -> u64 {
+pub const fn submul_nx1(lhs: &mut [u64], a: &[u64], b: u64) -> u64 {
     assume!(lhs.len() == a.len());
     let mut carry = 0;
     let mut borrow = false;
-    for i in 0..a.len() {
+    let mut i = 0;
+    while i < a.len() {
         // Compute product limbs
         let limb;
-        (limb, carry) = u128::muladd(a[i], b, carry).split();
+        (limb, carry) = DW::carrying_mul(a[i], b, carry).split();
 
         // Subtract
         (lhs[i], borrow) = borrowing_sub(lhs[i], limb, borrow);
+        i += 1;
     }
     borrow as u64 + carry
 }