+++ src/div.rs
@@ -5,7 +5,6 @@ impl<const BITS: usize, const LIMBS: usize> Uint<BITS, LIMBS> {
     /// Computes `self / rhs`, returning [`None`] if `rhs == 0`.
     #[inline]
     #[must_use]
-    #[allow(clippy::missing_const_for_fn)] // False positive
     pub fn checked_div(self, rhs: Self) -> Option<Self> {
         if rhs.is_zero() {
             return None;
@@ -16,7 +15,6 @@ impl<const BITS: usize, const LIMBS: usize> Uint<BITS, LIMBS> {
     /// Computes `self % rhs`, returning [`None`] if `rhs == 0`.
     #[inline]
     #[must_use]
-    #[allow(clippy::missing_const_for_fn)] // False positive
     pub fn checked_rem(self, rhs: Self) -> Option<Self> {
         if rhs.is_zero() {
             return None;